requirements.txt
Flask==2.1.1
validator==0.7.1
numpy==1.21.6
paddlepaddle==2.3.0
paddlespeech==1.0.1